Moved Code
Should be in the right place now so there's not duplicate calls, need to update yarn code to test which i do not know how to do
