+++ Assets/Scenes/FlowerMemoryScene.unity
@@ -129,6 +129,7 @@ GameObject:
   m_Component:
   - component: {fileID: 41747955}
   - component: {fileID: 41747958}
+  - component: {fileID: 41747959}
   m_Layer: 0
   m_Name: Manager
   m_TagString: Untagged
@@ -179,6 +180,20 @@ MonoBehaviour:
   flowerSprites:
   - {fileID: 857312311372029537, guid: 22b0d6ccf74eac744806de97c2df2787, type: 3}
   - {fileID: 599749539538422185, guid: 62c29300dc0fedc4286ef10ed0845fe1, type: 3}
+--- !u!114 &41747959
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 41747954}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 67637868ea98b044f925bdef5cb81cdf, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  dialogueRunner: {fileID: 492172087}
+  characterPortrait: {fileID: 1340726477}
 --- !u!1 &88269609
 GameObject:
   m_ObjectHideFlags: 0
