+++ Assets/Scripts/Minigames/DialogueMinigame.cs
@@ -11,17 +11,9 @@ public class DialogueMinigame : MinigameBehavior
     [SerializeField]
     private string startNode;
 
-    //Darken/Lighten Testing
-    [SerializeField]
-    private GameObject characterPortrait;
-
     public override void StartMinigame()
     {
         StartCoroutine(StartMinigameCoroutine());
-
-        //Darken/Lighten Testing
-        dialogueRunner.AddCommandHandler("lighten", Lighten);
-        dialogueRunner.AddCommandHandler("darken", Darken);
     }
 
     private IEnumerator StartMinigameCoroutine()
@@ -34,17 +26,4 @@ public class DialogueMinigame : MinigameBehavior
     {
     }
 
-    //Darken/Lighten Testing
-    private void Darken()
-    {
-        characterPortrait.GetComponent<SpriteRenderer>().color = new Color(194, 194, 194);
-        Debug.Log("Darken!");
-    }
-
-    //Darken/Lighten Testing
-    private void Lighten()
-    {
-        characterPortrait.GetComponent<SpriteRenderer>().color = Color.white;
-        Debug.Log("Lighten!");
-    }
 }
